+++ src/main/java/org/game/objects/strategy/attack/RandomAttackStrategy.java
@@ -5,11 +5,11 @@ import org.game.actions.AttackAction;
 import java.util.Random;
 
 public class RandomAttackStrategy implements AttackStrategy {
-    final double maxChanceAttack = 5.0;
+    final int maxChanceAttack = 5;
     Random random = new Random();
 
     @Override
     public AttackAction createAttack(String sender, String receiver) {
-        return new AttackAction(sender, receiver, random.nextDouble(maxChanceAttack));
+        return new AttackAction(sender, receiver, random.nextInt(maxChanceAttack));
     }
 }
